js1/requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cchardet==2.1.7
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
lxml==4.6.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
PyExecJS==1.5.1
Pygments==2.7.3
requests==2.25.1
six==1.15.0
soupsieve==2.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
yapf==0.30.0
